Merge in DOCTOOLS/documentation-portal from fix-pdf-copy-for-offline to master

* commit 'e917aed326afc051061f4ce333aadee9e25cce9c':
  Fix pdf copy for offline package to test for pdf first

+++ .teamcity/settings.kts
@@ -4349,7 +4349,7 @@ object GwBuildSteps {
                 #!/bin/bash
                 set -xe
                 
-                cp -avR "${onlineOutputPath}/pdf" "$offlineOutputPath"
+                [[ -e ${onlineOutputPath}/pdf/*.pdf ]] && cp -avR "${onlineOutputPath}/pdf" "$offlineOutputPath"
                 """.trimIndent()
         }
     }
